+++ net-inserter/src/component/mod.rs
@@ -1,1 +1,4 @@
+pub mod dispatcher;
 pub mod inserter;
+pub mod pcap_file_inserter;
+pub mod network;